style(*/create.php): improve form layout and add validation
The changes in both files improve the layout of the forms by centering them and adding more space between the elements. Additionally, validation has been added to the forms to ensure that the user enters valid data. The validation is done using JavaScript and the `novalidate` attribute has been added to the form to disable the browser's default validation.